Remove social functions
